nmad:   -- use 32 bits seq numbers by default
git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@27207 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ nmad/include/nm_core_interface.h
@@ -144,7 +144,7 @@ typedef uint16_t nm_req_flag_t;
 #define NM_FLAG_UNPACK_MATCHED       ((nm_req_flag_t)0x8000)
 
 /** Sequence number */
-typedef uint16_t nm_seq_t;
+typedef uint32_t nm_seq_t;
 
 /* ** tags ************************************************* */
